use pbuilder to build debianized packages
to avoid duplication of elbeprojects, once in the image once for the
pbuilder a new recipe 'elbeproject' is introduced. This is used to
manage the elbeproject and is the location for the source.xml and
de eproject file.

avoiding having 2 projects is important. Otherwise we need to sync
the project-local-repo from one project into another to have the
pbuilder packages available for image generation.

Signed-off-by: Manuel Traut <manut@linutronix.de>