classes source.xml.mako: adjust things for elbebase.bbclass
elbebase.bbclass calls the mako template expansion with bitbakes d
parameter as the single element of the namespace.

change source.xml.mako to get the necessary values via
d.getVar(), and also include the code that generated
the pkgs list into a <% %> block directly into the mako.

Also create <command> nodes into the finetuning from every line
in image_postprocess. Note that image_postprocess is a shell function.
in the calling bb file.

In this stage all lines of image_postprocess() are executed in
their private shell. This needs to be documented. And it might
be required that some statements are written to single lines.
(if statements, or variable usage)

Signed-off-by: Torben Hohn <torben.hohn@linutronix.de>
Signed-off-by: Manuel Traut <manut@linutronix.de>
Reviewed-by: Manuel Traut <manut@linutronix.de>