classes initvm: first implementation of initvm.bbclass
initvm.bbclass manages the initvm.

It allows to configure where the initvm resides, and uses lockfiles
to make sure, that only one is built. do_compile () is a python function
that calls the shell function do_initvm_create(), when the lockfiles
indicate that this is necessary.

it also defines some ELBE_INITVM_* variables, which are only
relevant for the initvm, and dont show up in elbebase.bbclass.

ELBE_BIN_WORKDIR and ELBE_INITVM_EXTRA_OPTS can be used to
use "elbe initvm create --devel" which is relevant during development.

ELBE_INITVM_URLLIST is expanded to debian urlnodes in the mako file.

Also add recipes-core/initvm/initvm.bb which uses this class.

Signed-off-by: Torben Hohn <torben.hohn@linutronix.de>
Signed-off-by: Manuel Traut <manut@linutronix.de>
Reviewed-by: Manuel Traut <manut@linutronix.de>