classes pbuilder: use git repack to remove references for source format git
The yocto fetcher clones a git repo into its download directory, and
then a shallow clone is used to clone the git repo into the working
directory.
When "elbe pbuilder build" uploads the build directory into the initvm
the referenced files are not uploaded. This prevents proper building
of the source package.

Fix that by calling "git repack -a" when source format git is detected.

Signed-off-by: Torben Hohn <torben.hohn@linutronix.de>
Signed-off-by: Manuel Traut <manut@linutronix.de>
Reviewed-by: Manuel Traut <manut@linutronix.de>