classes pbuilder: pass ELBE_PBUILDER_EXTRA_OPTS options to 'pbuilder build'
the qemu-user-static version used in elbe unstable results in a crash
when java is executed. this can be mitigated when multithreading is not
used.

Add ELBE_PBUILDER_EXTRA_OPTS variable, which allows passing arbitrary
options to the 'elbe pbuilder build' commandline.

This allows passing the --cpuset 1 option, which allows java builds to
succeed.

Signed-off-by: Torben Hohn <torben.hohn@linutronix.de>
Signed-off-by: Manuel Traut <manut@linutronix.de>
Reviewed-by: Manuel Traut <manut@linutronix.de>